+++ strato/VM/SolidVM/solid-vm/src/Blockchain/SolidVM/SM.hs
@@ -175,6 +175,7 @@ type MonadSM m = ( (Account `A.Alters` AddressState) m
                  , HasSelectX509CertDB m
                  , HasSelectX509FieldDB m
                  , A.Selectable Word256 ParentChainIds m
+                 , A.Selectable (Address,T.Text) X509CertificateField m
                  , HasRawStorageDB m
                  , HasMemAddressStateDB m
                  , HasMemRawStorageDB m
@@ -523,7 +524,7 @@ getVariableOfName name = do
       , maybeThis
       , maybeConstant
       --, maybeUserDefined
-      , unknownVariable ("getVariableOfName" ++ (show (currentContract currentCallInfo^.CC.storageDefs)) )name
+      , unknownVariable ("getVariableOfName " ++ (show (currentContract currentCallInfo^.CC.storageDefs)) )name
       ]
 
 getTypeOfName' :: SolidString -> CC.CodeCollection -> Typo