+++ strato/api/bloc/bloc2/src/Bloc/Server/Transaction.hs
@@ -11,6 +11,7 @@
 {-# LANGUAGE TypeOperators       #-}
 
 {-# OPTIONS -fno-warn-unused-top-binds #-}
+{-# OPTIONS -fno-warn-redundant-constraints #-}
 
 module Bloc.Server.Transaction (
   postBlocTransaction,
@@ -44,6 +45,7 @@ import           Data.Int                          (Int32)
 import           Data.List                         (partition, sortOn)
 import qualified Data.Vector                       as V
 
+import qualified Data.Map                          as M
 import           Data.Map.Strict                   (Map)
 import qualified Data.Map.Strict                   as Map
 import           Data.Maybe
@@ -82,6 +84,7 @@ import           BlockApps.Solidity.XabiContract
 import qualified BlockApps.Solidity.Xabi.Type      as Xabi
 import           Blockchain.Data.AddressStateDB
 import           Blockchain.Data.AlternateTransaction
+import           Blockchain.Data.CirrusDefs
 import           Blockchain.Data.DataDefs
 import           Blockchain.Data.Json             hiding (Contract)
 import           Blockchain.Data.TXOrigin
@@ -124,6 +127,7 @@ mergeTxParams inner outer = inner <|> outer
 txWorker :: ( MonadLogger m
             , A.Selectable Account Contract m
             , A.Selectable Account AddressState m
+            , A.Selectable Address Certificate m
             , HasCodeDB m
             , (Keccak256 `A.Selectable` SourceMap) m
             , HasBlocEnv m
@@ -477,6 +481,7 @@ postBlocTransactionUnsigned (Just jwt) cid (PostBlocTransactionRequest mAddr txL
 postBlocTransactionParallel :: ( MonadLogger m
                                , A.Selectable Account Contract m
                                , A.Selectable Account AddressState m
+                               , A.Selectable Address Certificate m
                                , HasCodeDB m
                                , (Keccak256 `A.Selectable` SourceMap) m
                                , HasBlocEnv m
@@ -502,6 +507,7 @@ postBlocTransactionParallel jwtToken b resolve queue c =
 postBlocTransaction :: ( MonadLogger m
                        , A.Selectable Account Contract m
                        , A.Selectable Account AddressState m
+                       , A.Selectable Address Certificate m
                        , HasCodeDB m
                        , (Keccak256 `A.Selectable` SourceMap) m
                        , HasBlocEnv m
@@ -518,6 +524,7 @@ postBlocTransaction = postBlocTransaction' (Don't CacheNonce)
 postBlocTransactionExternal :: ( MonadLogger m
                               , A.Selectable Account Contract m
                               , A.Selectable Account AddressState m
+                              , A.Selectable Address Certificate m
                               , HasCodeDB m
                               , (Keccak256 `A.Selectable` SourceMap) m
                               , HasBlocEnv m
@@ -535,6 +542,7 @@ postBlocTransactionExternal bearerToken = postBlocTransaction' (Don't CacheNonce
 postBlocTransaction' :: ( MonadLogger m
                         , A.Selectable Account Contract m
                         , A.Selectable Account AddressState m
+                        , A.Selectable Address Certificate m
                         , HasCodeDB m
                         , (Keccak256 `A.Selectable` SourceMap) m
                         , HasBlocEnv m
@@ -554,8 +562,16 @@ postBlocTransaction' cacheNonce mJwtToken chainId resolve (PostBlocTransactionRe
     Nothing -> throwIO $ UserError $ Text.pack "Did not find X-USER-ACCESS-TOKEN in the header"
     Just jwtToken -> do
       addr <- case mAddr of
-        Nothing -> fmap unAddress . blocVaultWrapper $  getKey (Just jwtToken) Nothing
+        Nothing -> fmap unAddress . blocVaultWrapper $ getKey (Just jwtToken) Nothing
         Just addr' -> return addr'
+      let err = CouldNotFind $ Text.concat
+                [ "postBlocTransaction': Couldn't find common name for user address "
+                , Text.pack $ formatAddressWithoutColor addr
+                ]
+      userCert <- maybe (throwIO err) pure =<<
+        A.select (A.Proxy @Certificate) addr
+      let userContractAddr = deriveAddressWithSalt Nothing (certificateCommonName userCert) Nothing "OrderedVals []"
+      $logInfoS "DEBUG" $ Text.pack $ show userContractAddr
       nonceMap <- getAccountNonce addr (S.singleton chainId)
       accountNonce <- case Map.lookup chainId nonceMap of
         Nothing -> pure $ 0
@@ -640,9 +656,9 @@ postBlocTransaction' cacheNonce mJwtToken chainId resolve (PostBlocTransactionRe
             p <- fromFunction x
             let bfp = FunctionParameters
                         addr
-                        (functionpayloadContractAddress p)
-                        (functionpayloadMethod p)
-                        (functionpayloadArgs p)
+                        userContractAddr
+                        "callContract"
+                        (M.fromList $ [("contractToCall",ArgString $ Text.pack $ show $ functionpayloadContractAddress p), ("functionName",ArgString $ functionpayloadMethod p), ("args", ArgArray $ V.fromList $ M.elems $ functionpayloadArgs p)])
                         (functionpayloadValue p)
                         (mergeTxParams (functionpayloadTxParams p) txParams)
                         (functionpayloadMetadata p)
@@ -654,7 +670,15 @@ postBlocTransaction' cacheNonce mJwtToken chainId resolve (PostBlocTransactionRe
             let bflp = FunctionListParameters
                         addr
                         (map (\(FunctionPayload a m r v x c md) ->
-                                MethodCall a m r (fromMaybe (Strung 0) v) (mergeTxParams x txParams) c md) p)
+                                MethodCall 
+                                  userContractAddr 
+                                  "callContract"  
+                                  (M.fromList $ [("contractToCall",ArgString $ Text.pack $ show a), ("functionName",ArgString m), ("args", ArgArray $ V.fromList $ M.elems r)])
+                                  (fromMaybe (Strung 0) v) 
+                                  (mergeTxParams x txParams) 
+                                  c 
+                                  md
+                              ) p)
                         chainId
                         resolve
             fmap BlocTxResult <$> postUsersContractMethodList' cacheNonce bflp jwtToken
@@ -1215,11 +1239,11 @@ getSolidityType _  Xabi.Account            = Right . SimpleType $ TypeAccount
 getSolidityType _ (Xabi.Struct _ name)     = Right $ TypeStruct name
 getSolidityType _ (Xabi.Enum _ name _)     = Right $ TypeEnum name
 getSolidityType _ (Xabi.Contract name)     = Right $ TypeContract name
-getSolidityType (ArgInt _) (Xabi.UnknownLabel _)  = Right $ SimpleType typeUInt -- since Enums are converted to Ints
+getSolidityType (ArgInt _) (Xabi.UnknownLabel _)     = Right $ SimpleType typeUInt -- since Enums are converted to Ints
 getSolidityType (ArgString _) (Xabi.UnknownLabel s)  = Right $ TypeEnum $ Text.pack s
 getSolidityType (ArgObject _) (Xabi.UnknownLabel s)  = Right $ TypeStruct $ Text.pack s --interpret an object strictly as a struct
 getSolidityType av (Xabi.UnknownLabel _)             = Left $ Text.pack $ "Expected a string, int, or object, but recieved: " ++ show av
-getSolidityType (ArgArray v) (Xabi.Array typ len)     =
+getSolidityType (ArgArray v) (Xabi.Array typ len)    =
   let arrType = case len of
         Just l -> TypeArrayFixed l
         Nothing -> TypeArrayDynamic
@@ -1230,6 +1254,7 @@ getSolidityType (ArgArray v) (Xabi.Array typ len)     =
 getSolidityType av (Xabi.Array _ _)          = Left $ Text.pack $ "Expected Array but got " ++ show av
 getSolidityType (ArgObject _) Xabi.Mapping{} = Right $ TypeStruct "s"
 getSolidityType av Xabi.Mapping{}            = Left $ Text.pack $ "Expected Object for Mapping type, but got " ++ show av
+getSolidityType _ Xabi.Variadic              = Right $ TypeVariadic
 
 
 getResultAndRespond :: ( A.Selectable Account AddressState m