+++ strato/CHANGELOG.md
@@ -22,12 +22,15 @@ so that they could be properly moved to their respective version's subsection.
 - `/transaction/unsigned` endpoint for generating raw transaction inputs
 - Bi-directional sync functionality
 - Mappings in SolidVM receive their own table in Cirrus
+- UserRegistry and User Contract on the genesis block
+- Connection to Cirrus Certificate table added in the API
 - Abstract contracts generate Cirrus tables
 - Derived contracts are inserted as rows in abstract tables
 - Support for imports from addresses in SolidVM
 - More lenient P2P disable times to prevent non-validators from being "locked out"
 ### Changed
 - `/compile` and `/transaction` endpoints use SolidVM compiler
+- POST `/transaction` calls redirected to the corresponding User contract
 ### Fixed
 - Error handle duplicate key violations in `code_ref` table
 ### Removed